Add parameter lists (attributes) to scopes
The whole scope (optional parameters + enclosed scope) is parsed under Parsec.try, because if the scope fails at the very end (e.g. because a constrained scope is followed by alphanum) then the parameter list needs to be reparsed.

At the same time its difficult to parse it directly as inlines, because "]" must interrupt any open scope, including extra "[". But this option is something that could be tried in the future.