Add inner library in .cabal file for debugging
It compiles the same modules than the public library, but using package parsec-free instead of parsec.parsec-free is a free-monad version of Parsec, with the same interface, useful for debugging parsers.

We add a new package to the cabal project: our own version of parsec-free (the original package is unmaintained), distributed separately.